Mac build ... builds and zips
Okay, the Mac build at this point finishes and has all the pieces
correctly where they ought to be.  However there is something wrong with
swt.  I can't get any version of it I've tried to work.  I can get the
app to start but the GUI doesn't work once you load a file.  And the
only way it starts like that from the app bundle is if the line adding
SWT to the classpath of the bundle is removed.

However it's not SWT's jar itself since if you run the jarfile by hand
and add the same SWT to the classpath, it doesn't immediately crash.  It
runs with the same GUI problem as if yu didn't specify it in the
classpath at all.

Something fishy is going on here.  But it builds, so perhaps others can
assist in figuring out the nature of the problem.  That's enough for me
to move on to the Linux build.