Add unix build conditionals
Previously AC built onm Windows or not Windows (assumed to be Mac).
It appears there was a means to build using gcj, but this is not
necessary, does not work anymore as configured, and isn't what we want
anyway.

Here's a first pass at improving that situation.  It's a bit ugly at
present because it requires a lot of markup to do something that should
be simple.  Still, I'd rather do this in steps.