Make multiplatform easier to read
This isn't significantly shorter than the first pass, but it's easier to
read and maintain.  To add a new platform to this (if anyone wants to do
that), just add sane default paths to ACBuild-default.properties, add a
conditional to detect your platform, and add an init-platform target
that does the configuration.

There's a bit cleaner solution in ant-contrib using the if task, but
that means adding to the dependencies and therefore the complexity of
building this thing.  I'm disinclined to do that.