Fix ant warning about includeantruntime
According to https://stackoverflow.com/questions/5103384, ant 1.8
introduced a "misfeature" of providing the ant runtime by default in the
java classpath.  This causes some problems with reproducible builds,
which is why later versions have a warning not to do it.

The solution is to make the inclusion of ant.jar in the classpath
explicit.  That's been done now.