Use the self-contained swt.jar from Eclipse
Rather than trying to use the Eclipse plugin version (which may never
actually work without building using Eclipse), I've changed the default
properties to prefer the all-in-one swt.jar packaged for separate
distribution.  That isn't going to make SWT magically start working (of
course not!), but it's got the best chance of us being able to MAKE it
work in the future.

I've elected to put it in /Users/Shared/Java which seems to be the
correct place to put such a thing, but elected to use an unversioned swt
symlink, set up like so:

/Users/Shared/Java/swt@ -> swt-4.7.1a-cocoa-macosx-x86_64

I'm using the current stable release version, as you see here.