Ensure macOS bundle starts the SWT GUI
In order to actually start the SWT-based GUI on macOS and have anything
start, you need -XstartOnMainThread.  This is a Cocoa limitation.  If
you have that (and we do now), you can pass the -swt argument and it
runs.  (More than I can say for Linux using the non-integrated SWT
provided by Debian at the moment…)

What doesn't work is the open dialog.  No files are selectable unless
you tell it to allow all files, and then .dsk files which the cmdline
jar file can handle fine are somehow an unsupported format.  Uh huh.

That much at least should not be too hard to fix.  Whether it works
after that is fixed, I cannot say.  What I can say is that the problem
does not exist if you don't pass -swt, however the lack of a proper
Swing GUI means it doesn't really matter because you can't do anything
with the files once you've opened them.