Make GUI request errors clean
If you don't request a specific GUI, we try SWT and then Swing.  Okay.
But if you do, we don't even try to see if it works—we just assume SWT
works.  If the user requested Swing, we just unconditionally spit out an
error message and try to start the GUI anyway.  Let's make the errors
conditional on the GUIs being unavailable.

Note, a Swing GUI is available.  It is just incomplete and doesn't work
very well.  :)