Support for future JavaFX GUI
This is a minor patch that adds detection (which will fail currently)
for a JavaFX GUI, when one is written.  It doesn't presently exist, but
when the ui/fx directory is created and populated with a JavaFX-based
GUI, it will start working and be used as the default choice.  Left
fallbacks for SWT and Swing, though of course the Swing fallback doesn't
have enough there to actually be useful yet.

I'm still teaching myself SWT and JavaFX, so this is all I've done with
it so far.  If anyone would like to contribute, please feel free.  :)