Do launchFXAppleCommander the FX way
In JavaFX applications, main is apparently not used unless the caller is
completely ignorant of JavaFX and how it works, but it's still
convention to provide it.  We could dynamically find our GUI and the
JavaFX Application class and use the latter to call the former.

Or y'know we can take advantage of the fact that main() can know what a
JavaFX Application is and just call it.

NOW we just need to write the JavaFX GUI!