Mod FilenameFilter to support JavaFX
cwa.storage.Disk.FileFilter was written for SWT, returning multiple
extensions as a single string with ; separators.  For JavaFX, we
actually want individual strings separately or in a list.

The only place these are initialized in cwa.storage.Disk anyway, so what
I've done is change the class to instatiate with separate strings and
join them on demand for SWT.  A new getExtensionList() method will
return them as-is for JavaFX.