Actually fix Linux build
At least on Debian, arch returns amd64 on Linux.  The moment I saw that
I realized that 32 bit wintel systems running Linux are going to return
everything from i386 to i686 at least.  Let's be proative about that if
we can.  :)

There's probably a plugin or module that does this behind the scenes,
but this is still much cleaner than what we had before.