Make my addition use correct whitespace
The downside of using <http://editorconfig.org/> is that I sometimes
don't look at what style is being used in a file, especially if the rest
of the project is using something else (somewhat) consistently.  Sorry
about that.  Recommend editorconfig though, especially for its ability
to automatically trim trailing whitespace, which is something good to do
in DVCS environment like git anyway.  Especially given that some editors
don't have a convenient end-user way to do it, and for others the way
looks something like…

    autocmd BufWritePre *.py :%s/\s\+$//e

D'oh.