Actually add something using JavaFX
This doesn't do much yet, but it doesn't do much using JavaFX, and I
think it's fair to suggest that it already does just about as much as
the Swing UI does.

Originally I tried to follow the Swing app model since it appeared that
it wanted to start out as a humble window and then grow into a full
thing once you opened something.  But really, there's some good reasons
why the SWT app doesn't do that—or at least there can be.

You probably don't want to merge this at the moment.  The Mac app is set
to start up using JavaFX rather than SWT by default.  Not that it
matters yet since the next thing I do to this tree will be to merge in
upstream's gradle build, and that pretty much nukes the macBundle target
for the time being anyway.  (Not a problem as until we beat this into
shape, it doesn't actually work anyway.)