Use the system menu bar (importan for macOS)
Currently our menu bar usage is lame and the SWT application doesn't
even bother.  It could be cool, but it currently isn't.  Hopefully that
will change.  For the moment though, it's at least on the top of the
screen if you are using a Mac where it belongs.

Of course the application menu is still "Java" and will be until we're
building an app bundle with Gradle, but that'll come in time.