Populate the directory tree pane in DiskWindow
I've decided (for now) not to try and break a DiskExplorerTab out of the
DiskWindow as was done with the SWT GUI just yet, mostly because it's
easier to write code when you don't have as many files to edit at once
(especially given java's really deeply nested trees!)

TreeViews in JavaFX always have one and only one root.  That's fine when
we only have one disk to display, but if we ever have two or more (the
only image I've got that I know has two disks is UniDOS's 800k image,
and that's not supported by AppleCommander at this time.)  When we have
that, we need a container tree root.

So what I did was create one unconditionally and hide it.  This gives
you a windows registry style interface with multiple roots, one for each
disk in the file you've opened.  That's fine.  But it still looks dumb
that the root of a ProDOS disk is not expanded.  I've resolved this by
checking for the single disk case an expanding the root directory in the
tree.

The toolbar and the right pane are still placeholders and there's only
the first tab.  This stuff should be broken out into separate files when
there's more of it done.